Moved validate_action_request() to manager
